New `Usb.Keycodes` & `Usb.Report` modules
Keycodes module adapted from QMK `keycode.h`, uses the same atoms for
each USB keycode ("normal" keys and also modifiers).

Provides an `is_modifier?/1` function (also introduces a `modifier?`
guard clause) and a `value/2` function that receives the atom
`keycode` (e.g. `:kc_a`) and the `state` and returns the USB keycode
that needs to be reflected on the USB input report.

The `Usb.Report` provides a `update_report_with_keys/2` and
`update_report/2` functions that allows us to update an input report to
send to the host.