Fix a bug detected with the Canon Typewriter pin combo + layout
It was a problem with the ordering of the pins. I was assuming the pins
were always ordered, but in the Canon Typewriter case we had them in
descending order.

Because of this, instead of sending the `Tab` key that should have been
sent when the `:kc_66` (A7 B7) is pressed, it was sending a `3` that
corresponds to `:kc_43` (A4 B3).

Why? Because the 8th pin (consider A7) counting from the end of the line
pins (i.e. `Ax`) was the A4 (AB, AA, A9, A8, A7, A6, A5, ->A4<-), and
the 7th pin (consider B7) from the end of the column pins (i.e. `Bx`)
was the B3 (BA, B9, B8, B7, B6, B5, B4, ->B3<-).

Sorting the pins solves the issue 😎