Simplify the handling of "transparent" keys
Instead of digging the actual keycode from the previous layer if a
keycode is "transparent" on a given layer, we simply call the
`keycode(kc_xy, N-1)` on the `keycode(kc_xy, N)` call for any layer N.

As an example, imagine that a layout has 3 layers (0, 1 and 2) and the
position `:kc_11` is mapped to `:kc_k` on layer 0, and is "transparent"
on layers 1 and 2:

- `keycode(:kc_11, 0)` returns `:kc_k`;
- `keycode(:kc_11, 1)` calls `keycode(:kc_11, 0)` because on layer 1
  position `:kc_11` is "transparent";
- `keycode(:kc_11, 2)` calls `keycode(:kc_11, 1)` because on layer 2
  position `:kc_11` is "transparent".

This simplifies a lot how we handle "transparent" positions (i.e.
positions mapped to `:____`).