Phoenix app working as expected by going to http://10.0.0.109/
We had to configure its `ElixirKeeb.UIWeb.Endpoint`, as we were doing for
the Phoenix app by itself.

Notice the `port`, `server: true` and especially the `check_origin:
true`, since without the latter the web socket connection from the
browser was getting a 403.

This isn't recommended and should be revisited.

Also, using the `Logs.PhoenixChannelBackend` log backend.