New PinMapper and Layout functions provided to its `use`rs
PinMapper provides `physical_matrix_kc_xy/0` and Layout provides
`keycode_by_physical_position/3`.

The PinMapper function returns the `:kc_xy` coordinates arranged by
their physical distribution on the keyboard. E.g.: On the Canon
Typewriter keyboard the Escape is connected to pins A6 B7, so the line
pin 5 and column pin 6 (pins are 0-index based), so the coordinate of
Escape is `:kc_56`.

Therefore, the `physical_matrix_kc_xy/0` function will return a list of
lists, where each list corresponds to a row. Escape is the first key of
the first row, so this function will return `:kc_56` on the first
position of the first list.

The Layout function lets one pass the physical row index, physical
column index and layer and will return the keycode in that position.
Continuing with the Escape example, calling
`keycode_by_physical_position(0, 0, 0)` will return the keycode in the
row 0, column 0 and layer 0, hence we will get `:kc_escape`.

This will be useful to return a string layout representation of the
keyboard layout to present on the web "dashboard" provided by the
`:elixir_keeb_ui`.