Activate PhoenixChannelBackend after 20s hopefully after Phoenix started
I've noticed that we can't immediately activate the
PhoenixChannelBackend Logger backend, since it relies on processes
started by Phoenix that take their time to start.

As such, we are for now waiting for 20s to then activate this Logger
backend. A better solution would be to check if Phoenix is available to
then activate this backend, and continue waiting if it's not the case.