Macro `KeycodeBehavior` always have a `:function` field
For each item of the `@macros` module entry, we create one `macro_x/1`
function in the module that is using the `ElixirKeeb.Layout` module.

The new `function` field is a handle to the corresponding `macro_x/1`
function that we created above.

If the `@macros` item is a list or a binary, we simply return a
`{macro_content, state}`, to replicate the behavior we already had for
string macros.